+++ src/lean_spec/subspecs/containers/signature.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from lean_spec.types import Bytes3100
+from lean_spec.types import Bytes3100, Uint64
 
 from ..xmss.containers import PublicKey
 from ..xmss.containers import Signature as XmssSignature
@@ -12,7 +12,7 @@ from ..xmss.interface import DEFAULT_SIGNATURE_SCHEME
 class Signature(Bytes3100):
     """Represents aggregated signature produced by the leanVM (SNARKs in the future)."""
 
-    def verify(self, public_key: PublicKey, epoch: int, message: bytes) -> bool:
+    def verify(self, public_key: PublicKey, epoch: Uint64, message: bytes) -> bool:
         """Verify the signature using XMSS verification algorithm."""
         try:
             scheme = DEFAULT_SIGNATURE_SCHEME
